requirements.txt
Flask
python-dotenv
chromadb
llama-index
llama-index-embeddings-google-genai
pydantic
typing-extensions
pandas
llama-index-vector-stores-chroma
llama-index-embeddings-google-genai